SUBMARINE-1208. Move files and refactor
