+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/rest/ModelVersionRestApi.java
@@ -46,8 +46,8 @@ import org.apache.submarine.server.model.database.service.ModelVersionService;
 
 
 import org.apache.submarine.server.model.database.service.ModelVersionTagService;
-import org.apache.submarine.server.response.JsonResponse;
 import org.apache.submarine.server.s3.Client;
+import org.apache.submarine.server.utils.JsonResponse;
 
 /**
  * Model version REST API v1.
