+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/rest/workbench/ParamRestApi.java
@@ -18,10 +18,10 @@
  */
 package org.apache.submarine.server.rest.workbench;
 
+import org.apache.submarine.server.utils.JsonResponse;
 import org.apache.submarine.server.rest.workbench.annotation.SubmarineApi;
 import org.apache.submarine.server.workbench.database.entity.ParamEntity;
 import org.apache.submarine.server.workbench.database.service.ParamService;
-import org.apache.submarine.server.response.JsonResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
