+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/rest/workbench/TeamRestApi.java
@@ -19,11 +19,12 @@
 package org.apache.submarine.server.rest.workbench;
 
 import com.github.pagehelper.PageInfo;
+
+import org.apache.submarine.server.utils.JsonResponse;
+import org.apache.submarine.server.utils.JsonResponse.ListResult;
 import org.apache.submarine.server.rest.workbench.annotation.SubmarineApi;
 import org.apache.submarine.server.workbench.database.entity.TeamEntity;
 import org.apache.submarine.server.workbench.database.service.TeamService;
-import org.apache.submarine.server.response.JsonResponse;
-import org.apache.submarine.server.response.JsonResponse.ListResult;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
