+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/utils/JsonResponse.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.submarine.server.response;
+package org.apache.submarine.server.utils;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.gson.Gson;
@@ -25,13 +25,7 @@ import com.google.gson.GsonBuilder;
 import com.google.gson.TypeAdapter;
 import org.apache.submarine.server.api.environment.EnvironmentId;
 import org.apache.submarine.server.api.experiment.ExperimentId;
-import org.apache.submarine.server.gson.EnvironmentIdDeserializer;
-import org.apache.submarine.server.gson.EnvironmentIdSerializer;
 import org.apache.submarine.server.api.notebook.NotebookId;
-import org.apache.submarine.server.gson.ExperimentIdDeserializer;
-import org.apache.submarine.server.gson.ExperimentIdSerializer;
-import org.apache.submarine.server.gson.NotebookIdDeserializer;
-import org.apache.submarine.server.gson.NotebookIdSerializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
