+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/rest/ExperimentRestApiTest.java
@@ -33,8 +33,8 @@ import org.apache.submarine.server.api.spec.ExperimentMeta;
 import org.apache.submarine.server.api.spec.ExperimentSpec;
 import org.apache.submarine.server.api.spec.KernelSpec;
 import org.apache.submarine.server.experiment.ExperimentManager;
-import org.apache.submarine.server.gson.ExperimentIdDeserializer;
-import org.apache.submarine.server.gson.ExperimentIdSerializer;
+import org.apache.submarine.server.utils.ExperimentIdDeserializer;
+import org.apache.submarine.server.utils.ExperimentIdSerializer;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.BeforeClass;
