+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/rest/RegisteredModelRestApiTest.java
@@ -33,10 +33,10 @@ import java.util.ArrayList;
 import java.util.List;
 import javax.ws.rs.core.Response;
 import org.apache.submarine.server.api.experiment.ExperimentId;
-import org.apache.submarine.server.gson.ExperimentIdDeserializer;
-import org.apache.submarine.server.gson.ExperimentIdSerializer;
 import org.apache.submarine.server.model.database.entities.RegisteredModelEntity;
 import org.apache.submarine.server.model.database.service.RegisteredModelService;
+import org.apache.submarine.server.utils.ExperimentIdDeserializer;
+import org.apache.submarine.server.utils.ExperimentIdSerializer;
 
 public class RegisteredModelRestApiTest {
   private final RegisteredModelService registeredModelService = new RegisteredModelService();
