+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/utils/JsonResponseTest.java
@@ -15,13 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.submarine.server.response;
+package org.apache.submarine.server.utils;
 
 import com.google.common.reflect.TypeToken;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+
+import org.apache.submarine.server.utils.JsonResponse.ListResult;
 import org.apache.submarine.server.workbench.database.entity.SysDictEntity;
-import org.apache.submarine.server.response.JsonResponse.ListResult;
 import org.junit.Test;
 
 import javax.ws.rs.core.Response;
