+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/workbench/database/utils/DictAnnotationTest.java
@@ -22,8 +22,9 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.internal.LinkedTreeMap;
 import com.google.gson.reflect.TypeToken;
-import org.apache.submarine.server.response.DictAnnotation;
-import org.apache.submarine.server.response.JsonResponse;
+
+import org.apache.submarine.server.utils.DictAnnotation;
+import org.apache.submarine.server.utils.JsonResponse;
 import org.apache.submarine.server.rest.workbench.CommonDataTest;
 import org.apache.submarine.server.rest.workbench.SysUserRestApi;
 import org.junit.Test;
