+++ submarine-test/test-k8s/src/test/java/org/apache/submarine/rest/EnvironmentManagerRestApiIT.java
@@ -27,9 +27,9 @@ import org.apache.commons.httpclient.methods.GetMethod;
 import org.apache.submarine.server.AbstractSubmarineServerTest;
 import org.apache.submarine.server.api.environment.Environment;
 import org.apache.submarine.server.api.environment.EnvironmentId;
-import org.apache.submarine.server.gson.EnvironmentIdDeserializer;
-import org.apache.submarine.server.gson.EnvironmentIdSerializer;
-import org.apache.submarine.server.response.JsonResponse;
+import org.apache.submarine.server.utils.EnvironmentIdDeserializer;
+import org.apache.submarine.server.utils.EnvironmentIdSerializer;
+import org.apache.submarine.server.utils.JsonResponse;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
