+++ submarine-test/test-k8s/src/test/java/org/apache/submarine/rest/ExperimentRestApiIT.java
@@ -37,13 +37,13 @@ import org.apache.submarine.server.AbstractSubmarineServerTest;
 import org.apache.submarine.server.api.environment.EnvironmentId;
 import org.apache.submarine.server.api.experiment.Experiment;
 import org.apache.submarine.server.api.experiment.ExperimentId;
-import org.apache.submarine.server.gson.EnvironmentIdDeserializer;
-import org.apache.submarine.server.gson.EnvironmentIdSerializer;
-import org.apache.submarine.server.gson.ExperimentIdDeserializer;
-import org.apache.submarine.server.gson.ExperimentIdSerializer;
 import org.apache.submarine.server.api.environment.Environment;
-import org.apache.submarine.server.response.JsonResponse;
 import org.apache.submarine.server.rest.RestConstants;
+import org.apache.submarine.server.utils.EnvironmentIdDeserializer;
+import org.apache.submarine.server.utils.EnvironmentIdSerializer;
+import org.apache.submarine.server.utils.ExperimentIdDeserializer;
+import org.apache.submarine.server.utils.ExperimentIdSerializer;
+import org.apache.submarine.server.utils.JsonResponse;
 import org.joda.time.DateTime;
 import org.junit.Assert;
 import org.junit.Before;
