+++ submarine-test/test-k8s/src/test/java/org/apache/submarine/rest/ExperimentTemplateManagerRestApiIT.java
@@ -34,10 +34,10 @@ import org.apache.submarine.server.api.experimenttemplate.ExperimentTemplate;
 import org.apache.submarine.server.api.experimenttemplate.ExperimentTemplateSubmit;
 import org.apache.submarine.server.api.spec.ExperimentSpec;
 import org.apache.submarine.server.api.spec.ExperimentTemplateSpec;
-import org.apache.submarine.server.gson.ExperimentIdDeserializer;
-import org.apache.submarine.server.gson.ExperimentIdSerializer;
-import org.apache.submarine.server.response.JsonResponse;
 import org.apache.submarine.server.rest.RestConstants;
+import org.apache.submarine.server.utils.ExperimentIdDeserializer;
+import org.apache.submarine.server.utils.ExperimentIdSerializer;
+import org.apache.submarine.server.utils.JsonResponse;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
